+++ libs/bpm-core/src/lib/notification/index.ts
@@ -1,5 +1,6 @@
 export * from './notification-delivery.service';
 export * from './notification-dispatcher.token';
+export * from './notification-observer.token';
 export * from './notification-preference.entity';
 export * from './notification.entity';
 export * from './notification.enums';